+++ src/pininterface/pinstatechange.cpp
@@ -7,7 +7,7 @@ Copyright Stichting Humanity4all 2015
 namespace pin_interface_n {
 
 PinStateChange::PinStateChange(
-        int pin_number,
+        uint8_t pin_number,
         switch_event_n::switch_state_t switch_state) {
     pinNumber = pin_number;
     switchState = switch_state;
