+++ src/printswitchstate.cpp
@@ -7,7 +7,7 @@ namespace debug_n {
 
 void print_switch_state(switch_event_n::switch_state_t switch_state[N_SWITCHES]) {
     char str[N_SWITCHES];
-    for (int i = 0; i < N_SWITCHES; i++) {
+    for (uint8_t i = 0; i < N_SWITCHES; i++) {
         if (switch_state[i] == switch_event_n::switch_state_t::PRESSED) {
             str[i] = '1';
         } else {
